new push and pull scripts

+++ pull.sh
@@ -1,6 +1,7 @@
 #!/bin/bash
-read -p server_ip
-scp  -P 8522 root@$server_ip:/root/saved_data/2020-*.zip .
+read -p "Server ip? " server_ip
+read -p "Server port? " server_port
+scp  -P $server_port root@$server_ip:/root/saved_data/2020-*.zip .
 ls | grep 2020 | grep -ne .zip
 
 # get user input, 
@@ -13,17 +14,20 @@ rm 2020-*.zip
 
 # unzip chosen file
 unzip ~/2020-*.zip -d ~/data_from_backup
-pushd ~/data_from_backup/home/*
-cp -R vimwiki ~/vimwiki
-popd
-
-pushd ~/data_from_backup/home/*/Documents
-buku -i buku_backup.db
-cp keep.com ~/keep.com # merge asta
-cp -R tools ~/tools/
-cp -R research ~/research/
-popd
-
-# Clean files
-rm -rf ~/data_from_backup/
+read -p "Automatic extract (yes/no)" userResponse
+if [ "$userResponse" = 'yes' ]; then
+	pushd ~/data_from_backup/home/*
+	cp -R vimwiki ~/vimwiki
+	popd
+	
+	pushd ~/data_from_backup/home/*/Documents
+	buku -i buku_backup.db
+	cp keep.com ~/keep.com # merge asta
+	cp -R tools ~/tools/
+	cp -R research ~/research/
+	popd
+	
+	# Clean files
+	rm -rf ~/data_from_backup/
+fi
 rm ~/2020-*.zip
